Autoscroll: measure time in same unit throughout program
Use the usec units that we use for the timer tick clock as the time base
for the whole program.

+++ worlds/local.h
@@ -1,7 +1,7 @@
 // Mark a position in time:
 struct mark {
 	struct coords coords;
-	float time;
+	int64_t time;
 };
 
 // Current state of world:
